pyproject.toml
[project]
name = "jetpp"
version = "0.1"
description = "Preprocessing for jet tagging"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "h5py==3.8.0",
    "numpy==1.24.*",
    "pre-commit==2.20.0",
    "pyyaml-include==1.3",
    "PyYAML==6.0",
    "rich==12.6.0",
    "scipy==1.9.3",
    "puma-hep"
]

[project.scripts]
preprocess = "jetpp.main:main"

[tool.setuptools]
packages = ["jetpp"]

[build-system]
requires = ["setuptools>=62"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
preview = "True"

[tool.ruff]
select = ["I", "E", "W", "F", "B", "UP", "ARG", "SIM", "TID", "RUF", "D2", "D3", "D4"]
ignore = ["D211", "D213", "RUF005"]
line-length = 100

[mypy]
ignore_missing_imports = "True"

[tool.pytest.ini_options]
log_cli_level = "debug"
filterwarnings = ["ignore::DeprecationWarning"]
